add: 生成grpc参数

+++ console/commands/route.go
@@ -13,11 +13,12 @@ import (
 // RouteCommand @Bean
 type RouteCommand struct{}
 
+// 简单gin处理模版
 var goMethodStr = `package {package}
 
 import ({import})
 
-// {action}  {doc}
+// {action}{doc}
 func (receiver *Controller) {action}(req *{paramAlias}.{param}, ctx http.Context) (*{returnAlias}.{return}, error) {
 	// TODO 这里写业务
 	return &{returnAlias}.{return}{}, nil
@@ -44,10 +45,60 @@ func (receiver *Controller) GinHandle{action}(ctx *gin.Context) {
 }
 `
 
+// http同时调用到grpc入口
+var goMethodStrForCallGrpc = `package {package}
+
+import ({import})
+
+// {action}  {doc}
+func (receiver *Controller) {action}(req *{paramAlias}.{param}, ctx http.Context) (*{returnAlias}.{return}, error) {
+	return myGrpc.NewHandle().{action}(context.Background(), req)
+}
+
+// GinHandle{action} gin原始路由处理
+// http.{method}({url})
+func (receiver *Controller) GinHandle{action}(ctx *gin.Context) {
+	req := &{paramAlias}.{param}{}
+	err := ctx.ShouldBind(req)
+	context := http.NewContext(ctx)
+	if err != nil {
+		context.Fail(err)
+		return
+	}
+	
+	resp, err := receiver.{action}(req, context)
+	if err != nil {
+		context.Fail(err)
+		return
+	}
+
+	context.Success(resp)
+}
+`
+
+// grpc入口
+var goGrpcFunc = `package {package}
+
+import ({import})
+
+// {action}{doc}
+// http.{method}({url})
+func (h Handle) {action}(ctx context.Context, req *{paramAlias}.{param}) (*{returnAlias}.{return}, error) {
+	// TODO 这里写业务
+	return &{returnAlias}.{return}{}, nil
+}
+`
+
 var goConStr = `package {package}
 
 // Controller @Bean
 type Controller struct {
+}`
+
+var goGrpcStr = `package {package}
+
+// Handle @Bean
+type Handle struct {
 }`
 
 func (RouteCommand) Configure() command.Configure {
@@ -70,6 +121,12 @@ func (RouteCommand) Configure() command.Configure {
 					Name:        "out_actions",
 					Description: "生成文件到指定目录",
 					Default:     "@root/app/http",
+				},
+			},
+			Has: []command.ArgParam{
+				{
+					Name:        "--grpc",
+					Description: "生成grpc文件处理模版",
 				},
 			},
 		},
@@ -109,7 +166,7 @@ func (RouteCommand) Execute(input command.Input) {
 								servers:     make([]parser.Service, 0),
 							}
 						}
-						genController(service, outHttp)
+						genController(service, outHttp, input.GetHas("--grpc"))
 						break
 					}
 				}
@@ -144,7 +201,7 @@ func (RouteCommand) Execute(input command.Input) {
 	_ = cmd.Run()
 }
 
-func genController(server parser.Service, out string) {
+func genController(server parser.Service, out string, genGrpc bool) {
 	page := server.Protoc.PackageName
 	out += "/" + page + "/" + parser.StringToSnake(server.Name)
 
@@ -167,21 +224,25 @@ func genController(server parser.Service, out string) {
 		}
 	}
 
-	methodStr := goMethodStr
 	gin := "github.com/gin-gonic/gin"
 	http := "github.com/go-home-admin/home/app/http"
 	imports := map[string]string{gin: gin, http: http}
 	goMod := getModModule()
+	methodStr := goMethodStr
 
+	grpcOut := strings.ReplaceAll(out, "/app/http", "/app/grpc")
+	if genGrpc {
+		methodStr = goMethodStrForCallGrpc
+		imports["myGrpc"] = goMod + "/app/grpc/handles/" + parser.StringToSnake(server.Name)
+		imports["context"] = "context"
+		// 生成grpc handle
+		genGrpcHandle(grpcOut, server)
+	}
 	for rName, rpc := range server.Rpc {
 		for _, options := range rpc.Opt {
 			for _, option := range options {
 				if strings.Index(option.Key, "http.") == 0 {
 					actionName := parser.StringToHump(rName)
-					actionFile := out + "/" + parser.StringToSnake(actionName) + "_action.go"
-					if parser.DirIsExist(actionFile) {
-						continue
-					}
 
 					serPage := goMod + "/generate/proto/" + server.Protoc.PackageName
 					imports[serPage] = serPage
@@ -218,7 +279,7 @@ func genController(server parser.Service, out string) {
 					method := option.Key[i+1:]
 					url := option.Val
 
-					for s, O := range map[string]string{
+					reps := map[string]string{
 						"{package}":         parser.StringToSnake(server.Name),
 						"{import}":          importsStr + "\n",
 						"{doc}":             rpc.Doc,
@@ -230,16 +291,48 @@ func genController(server parser.Service, out string) {
 						"{param}":           params,
 						"{returnAlias}":     returnAlias,
 						"{return}":          ret,
-					} {
-						str = strings.ReplaceAll(str, s, O)
 					}
 
-					err := os.WriteFile(out+"/"+parser.StringToSnake(actionName)+"_action.go", []byte(str), 0766)
-					if err != nil {
-						log.Fatal(err)
+					// 生成http
+					if !parser.DirIsExist(out + "/" + parser.StringToSnake(actionName) + "_action.go") {
+						for s, O := range reps {
+							str = strings.ReplaceAll(str, s, O)
+						}
+						err := os.WriteFile(out+"/"+parser.StringToSnake(actionName)+"_action.go", []byte(str), 0766)
+						if err != nil {
+							log.Fatal(err)
+						}
+					}
+
+					// 生成grpc
+					if genGrpc && !parser.DirIsExist(grpcOut+"/"+parser.StringToSnake(actionName)+"_action.go") {
+						reps["{import}"] = "\n\t\"context\"\n\t\"" + serPage + "\"\n"
+
+						goGrpcFunc2 := goGrpcFunc
+						for s, O := range reps {
+							goGrpcFunc2 = strings.ReplaceAll(goGrpcFunc2, s, O)
+						}
+						err := os.WriteFile(grpcOut+"/"+parser.StringToSnake(actionName)+"_action.go", []byte(goGrpcFunc2), 0766)
+						if err != nil {
+							log.Fatal(err)
+						}
 					}
 				}
 			}
+		}
+	}
+}
+
+func genGrpcHandle(out string, server parser.Service) {
+	if !parser.DirIsExist(out) {
+		_ = os.MkdirAll(out, 0760)
+	}
+
+	if !parser.DirIsExist(out + "/handle.go") {
+		conStr := strings.ReplaceAll(goGrpcStr, "{package}", parser.StringToSnake(server.Name))
+		err := os.WriteFile(out+"/handle.go", []byte(conStr), 0760)
+		if err != nil {
+			panic(err)
 		}
 	}
 }
